Added the menu for the log out capability. Actual log out functionality to follow later.
Added a click handler for the FAB which opens a dialog where the user
can enter the name of a new task group that they want to create. Actual
task creation functionality to follow later.

Filled the task group list with dummy task groups for testing. These
dummy values shall be removed once the database implementation is
complete.

Added a click handler for clicking on a task group within the list.
Right now clicking on any task group launches the task group activity.
Later on clicking on a task group will actually launch the task group
activity for the specific task group.