Added margins to the root element so that there is space between each card.
Added specific fonts and colors to the text views.

Added the actual icons that will be used within each image view.

The views in the icon row are chained.

Removed unneeded top and bottom margins for views within the icon row
since the views are in a chain, they will share the top and bottom
margin values.