Added a top-level to allow switching between to different root elements based on the position within the recycler view.
Assigned an id to each root element so that they are easy to hide when
switching roots.

The root element must have a predefined with or the recycler view
behaves strangely.

Apply a small bit of margin to the bottom tool bar.

Changed the color of the tool bar to a more appropriate value.

Changed the margins and font of the test in the toolbar.

Make sure the text in the tool bar and the menu button are in a chain.

Make sure that the recycler view has the same width as its parent or
strange things happen with the height of the list.

Constrain the recycler view height to a specific value. Again strange
things happen to the height if the value is lower.

Give the button a new font, color and background color.

Added an sibling root layout for the add task list button.