taskListList uses the TaskListModelContainer type not String.
Use the task group that was sent via the intent.

Use the task group list that was sent via the intent.

Use the getStartIntent pattern.

Handle configuration change.

Set the action bar title to the name of the task group.

Ensure the choose background dialog has the current background color
selected when initially displayed.

Handle card clicks from a task list.

Handle the add card button on a task list.

Implemented the add task list button.

Implemented the move task list menu.

Use the getStartIntent pattern to create the intent that launches the
CardDetailActivity.

Added a stub to fetch remote data.

Adjust the speed at which the task group activity scrolls while dragging
a card so that the scrolling is less jerky.

Ensure the task group activity scrolls when a card is dragged need its
horizontal bounds.