Save the index of the "task list" to the "task group" "task list" instead of "true".
Implemented task list move functionality.

Tell Firebase to order the task lists that we fetch by their indexes.

Tell Firebase to order the task list cards that we fetch by their
indexes.

Fetch the due date for a task list card from the database.

Only call notifyDataSetChanged on a task list adapter if we are sure it
exists.