Check for network connectivity before fetching remote data during the initialization phase.
Check for network connectivity before allowing a task group activity
to be launched.

Check for network connectivity before allowing a new task group to be
created.

Handle a user not being signed in when a user is attempting to create a
task group. Although this should never happen since we check a user's
log in status during initialization.

Show an error message if there is no network connectivity and this
activity has no task groups already displayed.