Handle the case where the task group activity is being initialized for the first time by coming from the card detail activity (its child). This can happen when the card detail activity is started via the app widget.
Allow the broadcast receiver to initialize data fetching based on
network connectivity.

Handle the case where a change is made in the card detail activity that
needs to be reflected in its parent, this activity.

Clear out the attachment data on a configuration changed so that we do
not hit the Android intent size limit.

Ensure we have network connectivity before allowing the following
situations.

1) Background selection.

2) Adding a task list card to a task list.

3) Adding a task list.

4) Moving a task list.

5) Clicking on a task list card.

Fetch checklist and attachment info so the we can set the appropriate
icons on a task list card.