Added an error message that will be displayed when there is no network connectivity and there are no task list groups currently visible.
Changed the visibility of the error messages from "gone" to "invisible"
in order to overcome a issue (bug maybe) that happens occurs when a
widget in this case a text view is initially gone then set visible. The
text view shall be visible as confirmed by a check to the view's visible
state. However, the view's height will be 0! Effectively rendering the
view invisible.