Update the intent that is used to launch this activity to include the task group model and the task group model list. This is needed to add in a "back stack" when this activity is launched via the app widget.
Attachment data is longer saved during a configuration change in order
to overcome a limit in the Android platform on the size of parcelable
data.

Override onResume to handle coming back to the activity when selecting
a new attachment.

Check for network connectivity in various cases where the database needs
to be updated following a user changing the data.

Fix a bug where the entire database was deleted when a due date was
removed! Oops. :-(

Update the widget when a due date is removed or completed.

Check for null when parsing a date (Although null should never occur).

Update the database when a checklist is deleted.

Check to make sure we are dealing with an image attachment before
setting the bound flag. The bound flag is only valid for images.

Updated the attachment model so that attachments have a unique ID.

Update the database when an attachment is added/removed.

Save the initial text for edit text views when the user makes an edit.
We can then compare the initial state with the current state to
determine if a database update is required.