Ensure checklist items get a unique ID.
Update the database when there is a change to a checklist item.

Handle an edge case when the user deletes a checklist item and the item
receives a focus lost event.

Check for network connectivity before allowing the user to make changes
to an item that would result in a database update.