Added a back stack so that the card detail activity can go back to its parents when the back button is pressed. The remote views factory is responsible for added the extra data for just the card detail activity using a fill-in intent.
Added a "workaround" (I guess) for a case when the user logs out of the
app but instead of log in button being displayed, the empty widget is
displayed.

Added a helper function to start the fetch IntentService.

Added a helper function to start the delete IntentService.

Added a helper function to start the update IntentService.

Override onReceive so that the various IntentServices can communicate
status updates to the widget.